+++ packages/cloud-cognitive/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.77.3](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-cloud-cognitive@0.77.2...@carbon/ibm-cloud-cognitive@0.77.3) (2021-08-26)
+
+
+### Bug Fixes
+
+* **SidePanel:** fix label text layout issues on scroll animation ([#1203](https://github.com/carbon-design-system/ibm-cloud-cognitive/issues/1203)) ([d1c5570](https://github.com/carbon-design-system/ibm-cloud-cognitive/commit/d1c5570a56eea58157b8ad2e8faca225787f8479))
+
+
+
+
+
+## [0.77.2](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-cloud-cognitive@0.77.1...@carbon/ibm-cloud-cognitive@0.77.2) (2021-08-26)
+
+**Note:** Version bump only for package @carbon/ibm-cloud-cognitive
+
+
+
+
+
 ## [0.77.1](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-cloud-cognitive@0.77.0...@carbon/ibm-cloud-cognitive@0.77.1) (2021-08-26)